// frontend/src/api_gen/models/country-daily-wage-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CountryDailyWageDTO
 */
export interface CountryDailyWageDTO {
    /**
     * 
     * @type {string}
     * @memberof CountryDailyWageDTO
     */
    'countryName'?: string;
    /**
     * 
     * @type {string}
     * @memberof CountryDailyWageDTO
     */
    'countryCode'?: string;
    /**
     * 
     * @type {string}
     * @memberof CountryDailyWageDTO
     */
    'continentName'?: string;
    /**
     * 
     * @type {number}
     * @memberof CountryDailyWageDTO
     */
    'eurDailyWage'?: number;
}